requirements.txt
aiohttp==3.9.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
alembic==1.11.3
    # via dagster
altair==5.3.0
    # via streamlit
aniso8601==9.0.1
    # via graphene
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
authlib==1.3.0
    # via weaviate-client
backoff==2.2.1
    # via gql
beautifulsoup4==4.12.3
    # via -r requirements.in
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   dagster
    #   dagster-webserver
    #   nltk
    #   streamlit
    #   uvicorn
coloredlogs==14.0
    # via dagster
contourpy==1.2.1
    # via matplotlib
croniter==1.4.1
    # via dagster
cryptography==42.0.7
    # via authlib
cycler==0.12.1
    # via matplotlib
dagster==1.7.7
    # via
    #   -r requirements.in
    #   dagster-graphql
    #   dagster-webserver
dagster-graphql==1.7.7
    # via dagster-webserver
dagster-pipes==1.7.7
    # via dagster
dagster-webserver==1.7.7
    # via -r requirements.in
dataclasses-json==0.6.6
    # via langchain-community
distro==1.9.0
    # via openai
docstring-parser==0.15
    # via dagster
filelock==3.14.0
    # via dagster
fonttools==4.53.0
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.6.0
    # via universal-pathlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
gql[requests]==3.4.1
    # via dagster-graphql
graphene==3.3
    # via dagster-graphql
graphql-core==3.2.3
    # via
    #   gql
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
greenlet==2.0.2
    # via sqlalchemy
grpcio==1.57.0
    # via
    #   dagster
    #   grpcio-health-checking
    #   grpcio-tools
    #   weaviate-client
grpcio-health-checking==1.57.0
    # via
    #   dagster
    #   weaviate-client
grpcio-tools==1.57.0
    # via weaviate-client
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.27.0
    # via
    #   openai
    #   weaviate-client
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.2
    # via
    #   altair
    #   dagster
    #   pydeck
joblib==1.4.2
    # via
    #   nltk
    #   scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
langchain==0.2.1
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.1
    # via -r requirements.in
langchain-core==0.2.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.1.7
    # via -r requirements.in
langchain-text-splitters==0.2.0
    # via langchain
langsmith==0.1.63
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
mako==1.2.4
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
marshmallow==3.21.2
    # via dataclasses-json
matplotlib==3.9.0
    # via
    #   -r requirements.in
    #   wordcloud
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   langchain
    #   langchain-community
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   streamlit
    #   wordcloud
openai==1.30.4
    # via langchain-openai
orjson==3.10.3
    # via langsmith
packaging==23.2
    # via
    #   altair
    #   dagster
    #   langchain-core
    #   marshmallow
    #   matplotlib
    #   streamlit
pandas==2.2.2
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pendulum==2.1.2
    # via dagster
pillow==10.3.0
    # via
    #   matplotlib
    #   streamlit
    #   wordcloud
protobuf==4.24.2
    # via
    #   dagster
    #   grpcio-health-checking
    #   grpcio-tools
    #   streamlit
pyarrow==16.1.0
    # via streamlit
pycparser==2.22
    # via cffi
pydantic==2.7.1
    # via
    #   dagster
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   weaviate-client
pydantic-core==2.18.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   croniter
    #   dagster
    #   matplotlib
    #   pandas
    #   pendulum
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   dagster
    #   uvicorn
pytz==2023.3
    # via
    #   dagster
    #   pandas
pytzdata==2020.1
    # via pendulum
pyyaml==6.0.1
    # via
    #   dagster
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via
    #   nltk
    #   tiktoken
requests==2.31.0
    # via
    #   -r requirements.in
    #   dagster
    #   dagster-graphql
    #   gql
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
    #   weaviate-client
requests-toolbelt==0.10.1
    # via gql
rich==13.7.1
    # via
    #   dagster
    #   streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.5.0
    # via -r requirements.in
scipy==1.13.1
    # via scikit-learn
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.20
    # via
    #   alembic
    #   dagster
    #   langchain
    #   langchain-community
starlette==0.31.1
    # via
    #   dagster-graphql
    #   dagster-webserver
streamlit==1.35.0
    # via -r requirements.in
structlog==24.2.0
    # via dagster
tabulate==0.9.0
    # via dagster
tenacity==8.3.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
threadpoolctl==3.5.0
    # via scikit-learn
tiktoken==0.7.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tomli==2.0.1
    # via dagster
toolz==0.12.1
    # via altair
toposort==1.10
    # via dagster
tornado==6.4
    # via streamlit
tqdm==4.66.1
    # via
    #   dagster
    #   nltk
    #   openai
typing-extensions==4.7.1
    # via
    #   alembic
    #   dagster
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
universal-pathlib==0.1.2
    # via dagster
urllib3==1.26.16
    # via
    #   gql
    #   requests
uvicorn[standard]==0.23.2
    # via dagster-webserver
uvloop==0.17.0
    # via uvicorn
validators==0.28.1
    # via weaviate-client
watchdog==3.0.0
    # via
    #   dagster
    #   streamlit
watchfiles==0.20.0
    # via uvicorn
weaviate-client==4.6.3
    # via -r requirements.in
websockets==11.0.3
    # via uvicorn
wordcloud==1.9.3
    # via -r requirements.in
yarl==1.9.2
    # via
    #   aiohttp
    #   gql

# The following packages are considered to be unsafe in a requirements file:
# setuptools
